Resolved a few bugs and the Organizer Notification page is fucntional now
